L3-24: a few file moves and simplier db connection.
