+++ package.json
@@ -8,7 +8,7 @@
     "build:tsoa": "tsoa spec-and-routes",
     "build": "tsoa spec-and-routes && tsc",
     "start": "node build/index.js",
-    "dev": "npm run build && concurrently \"npx tsc --watch\" \"nodemon -q build/index.js | pino-colada\"",
+    "dev": "npm run build && NODE_ENV=dev concurrently \"npx tsc --watch\" \"nodemon -q build/index.js | pino-colada\"",
     "test:unit": "NODE_ENV=test ./node_modules/.bin/mocha -r ts-node/register ./src/**/*.test.ts",
     "test:integration": "NODE_ENV=test ./node_modules/.bin/mocha --timeout 30000 -r ts-node/register ./integration/*.test.ts"
   },
